Move Done

+++ object.cpp
@@ -5,7 +5,7 @@ Object::Object(QObject *parent) : QObject(parent)
 
 }
 
-Object::Object(int x, int y, QColor color):x{x},y{y},
+Object::Object(float x,float y,QColor color):x{x},y{y},
     m_color{color},m_brush{QBrush(color)}
 {
